round1b/requirements.txt
langchain-community==0.3.13
pypdf==5.1.0
chromadb==0.5.23
sentence-transformers==3.3.1
llama-cpp-python==0.3.2
numpy==1.24.3
pandas==2.0.3